# PERF: Timestamp construction with tz and string offset